Example revision (#3)
* Adding library.json

* update library.json

* deleting a duplicate in library.json

* Moving files in src, and deleting src for SPIN examples

* Addind details and schema to DC_DC twist examples, correcting main boost example

* adding a general readme file

* Correcting comment error in multiple PWM example

* Correct typo in library.json

* adding a appconfig for test

* adding app.conf files where needed

* deleting TWIST mention in SPIN examples

* deleting unused variables warning

adding macro #ifdef where needed

* Rewrtting loop application task

 - deleting while loop and k_msleep in twist examples
- Now more coherent with current core version

* adding authors in examples and library.json

* reorganizing library.json

Putting easier spin examples first in the list
