+++ README.md
@@ -1,0 +1,38 @@
+Welcome to the owntech examples repository.
+
+Here you can find ready to be used example for [spin](https://github.com/owntech-foundation/SPIN) and [twist](https://github.com/owntech-foundation/TWIST).
+
+# Example for TWIST
+
+- DC DC topology examples
+    - [Buck voltage mode](TWIST/DC_DC/buck_voltage_mode/README.md)
+    - [Buck current mode](TWIST/DC_DC/buck_current_mode/README.md)
+    - [Boost voltage mode](TWIST/DC_DC/boost_voltage_mode/README.md)
+    - [Interleaved Buck](TWIST/DC_DC/interleaved/README.md)
+
+- Microgrid examples
+    - [AC client server](TWIST/Microgrid/AC_client_server/README.md)
+    - [AC peer to peer](TWIST/Microgrid/AC_peer_to_peer/README.md)
+    - [DC client server](TWIST/Microgrid/DC_client_server/README.md)
+    - [DC peer to peer](TWIST/Microgrid/DC_droop/README.md)
+    - [DC droop](TWIST/Microgrid/DC_peer_to_peer/README.md)
+
+# Example for SPIN
+
+- ADC use examples
+    - [Software trigerred ADC](SPIN/ADC/adc_software_trigger/README.md)
+    - [HRTIM trigerred ADC](SPIN/ADC/adc_hrtim_trigger/README.md)
+
+- DAC use examples
+    - [Signal generation](SPIN/DAC/signal_generation/README.md)
+
+- LED use examples
+    - [Blinky](SPIN/LED/blinky/README.md)
+
+- PWM use examples
+    - [PWM duty cycle control](SPIN/PWM/duty_cycle_setting/README.md)
+    - [PWM phase shift control](SPIN/PWM/phase_shift/README.md)
+    - [Multiple PWM operation](SPIN/PWM/multiple_pwm/README.md)
+
+- TIMER
+    - [Incremental encoder](SPIN/TIMER/incremental_encoder/README.md)
