+++ SPIN/LED/blinky/main.cpp
@@ -25,6 +25,7 @@
  *
  * @author Clément Foucher <clement.foucher@laas.fr>
  * @author Luiz Villa <luiz.villa@laas.fr>
+ * @author Ayoub Farah Hassan <ayoub.farah-hassan@laas.fr>
  */
 
 //--------------OWNTECH APIs----------------------------------
@@ -41,7 +42,7 @@ void loop_background_task();   // Code to be executed in the background task
 void loop_critical_task();     // Code to be executed in real time in the critical task
 
 //--------------USER VARIABLES DECLARATIONS-------------------
-static float32_t adc_value;
+
 
 
 //--------------SETUP FUNCTIONS-------------------------------
@@ -55,19 +56,15 @@ static float32_t adc_value;
 void setup_routine()
 {
     // Setup the hardware first
-    spin.version.setBoardVersion(TWIST_v_1_1_2);
-
-    spin.adc.configureTriggerSource(2, software); // ADC 2 configured in software mode
-
-    data.enableAcquisition(2, 35); // ADC 2 enabled
+    spin.version.setBoardVersion(SPIN_v_1_0);
 
     // Then declare tasks
     uint32_t background_task_number = task.createBackground(loop_background_task);
-    task.createCritical(loop_critical_task, 100); // Uncomment if you use the critical task
+    //task.createCritical(loop_critical_task, 500); // Uncomment if you use the critical task
 
     // Finally, start tasks
     task.startBackground(background_task_number);
-    task.startCritical(); // Uncomment if you use the critical task
+    //task.startCritical(); // Uncomment if you use the critical task
 }
 
 //--------------LOOP FUNCTIONS--------------------------------
@@ -80,10 +77,10 @@ void setup_routine()
 void loop_background_task()
 {
     // Task content
-    printk("%f \n", adc_value);
+    spin.led.toggle();
 
     // Pause between two runs of the task
-    task.suspendBackgroundMs(100);
+    task.suspendBackgroundMs(1000);
 }
 
 /**
@@ -94,8 +91,7 @@ void loop_background_task()
  */
 void loop_critical_task()
 {
-    data.triggerAcquisition(2);
-    adc_value = data.getLatest(2, 35);
+
 }
 
 /**
