+++ SPIN/PWM/duty_cycle_setting/main.cpp
@@ -25,6 +25,7 @@
  *
  * @author Clément Foucher <clement.foucher@laas.fr>
  * @author Luiz Villa <luiz.villa@laas.fr>
+ * @author Ayoub Farah Hassan <ayoub.farah-hassan@laas.fr>
  */
 
 //--------------OWNTECH APIs----------------------------------
@@ -32,17 +33,32 @@
 #include "TaskAPI.h"
 #include "TwistAPI.h"
 #include "SpinAPI.h"
+#include "opalib_control_pid.h"
+
+#include "zephyr/console/console.h"
 
 //--------------SETUP FUNCTIONS DECLARATION-------------------
 void setup_routine(); // Setups the hardware and software of the system
 
 //--------------LOOP FUNCTIONS DECLARATION--------------------
-void loop_background_task();   // Code to be executed in the background task
+void loop_communication_task(); // code to be executed in the slow communication task
+void loop_application_task();   // Code to be executed in the background task
 void loop_critical_task();     // Code to be executed in real time in the critical task
 
 //--------------USER VARIABLES DECLARATIONS-------------------
-static float32_t adc_value;
+uint8_t received_serial_char;
 
+float32_t duty_cycle = 0.3;
+
+//---------------------------------------------------------------
+
+enum serial_interface_menu_mode // LIST OF POSSIBLE MODES FOR THE OWNTECH CONVERTER
+{
+    IDLEMODE = 0,
+    POWERMODE
+};
+
+uint8_t mode = IDLEMODE;
 
 //--------------SETUP FUNCTIONS-------------------------------
 
@@ -55,49 +71,67 @@ static float32_t adc_value;
 void setup_routine()
 {
     // Setup the hardware first
-    spin.version.setBoardVersion(TWIST_v_1_1_2);
+    spin.version.setBoardVersion(SPIN_v_1_0);
 
     spin.pwm.setFrequency(200000); // Set frequency of pwm
-
-    spin.pwm.setModulation(PWMA, UpDwn);
-    spin.pwm.setAdcEdgeTrigger(PWMA, EdgeTrigger_up);
-
     spin.pwm.initUnit(PWMA); // timer initialization
-
-    // Setting trigger for ADC
-    spin.pwm.setAdcTrigger(PWMA, ADCTRIG_1);
-    spin.pwm.setAdcTriggerInstant(PWMA, 0.06);
-    spin.pwm.enableAdcTrigger(PWMA);
 
     spin.pwm.startDualOutput(PWMA); // Start PWM
 
-    spin.adc.configureTriggerSource(2, hrtim_ev1); // ADC 2 configured to be triggered by the PWM
-
-    data.enableAcquisition(2, 35); // ADC 2 enabled
-
     // Then declare tasks
-    uint32_t background_task_number = task.createBackground(loop_background_task);
+    uint32_t app_task_number = task.createBackground(loop_application_task);
+    uint32_t com_task_number = task.createBackground(loop_communication_task);
     task.createCritical(loop_critical_task, 100); // Uncomment if you use the critical task
 
     // Finally, start tasks
-    task.startBackground(background_task_number);
+    task.startBackground(app_task_number);
+    task.startBackground(com_task_number);
     task.startCritical(); // Uncomment if you use the critical task
 }
 
 //--------------LOOP FUNCTIONS--------------------------------
+
+void loop_communication_task()
+{
+    while (1)
+    {
+        received_serial_char = console_getchar();
+        switch (received_serial_char)
+        {
+        case 'h':
+            //----------SERIAL INTERFACE MENU-----------------------
+            printk(" ________________________________________\n");
+            printk("|     ------- MENU ---------             |\n");
+            printk("|     press u : duty cycle UP            |\n");
+            printk("|     press d : duty cycle DOWN          |\n");
+            printk("|________________________________________|\n\n");
+            //------------------------------------------------------
+            break;
+        case 'u':
+            duty_cycle += 0.05;
+            break;
+        case 'd':
+            duty_cycle -= 0.05;
+            break;
+        default:
+            break;
+        }
+    }
+}
 
 /**
  * This is the code loop of the background task
  * It is executed second as defined by it suspend task in its last line.
  * You can use it to execute slow code such as state-machines.
  */
-void loop_background_task()
+void loop_application_task()
 {
     // Task content
-    printk("%f \n", adc_value);
+    printk("%f\n", duty_cycle);
 
     // Pause between two runs of the task
-    task.suspendBackgroundMs(100);
+    task.suspendBackgroundMs(1000);
+
 }
 
 /**
@@ -108,7 +142,7 @@ void loop_background_task()
  */
 void loop_critical_task()
 {
-    adc_value = data.getLatest(2, 35);
+        spin.pwm.setDutyCycle(PWMA, duty_cycle);
 }
 
 /**
