+++ SPIN/TIMER/incremental_encoder/main.cpp
@@ -25,6 +25,7 @@
  *
  * @author Clément Foucher <clement.foucher@laas.fr>
  * @author Luiz Villa <luiz.villa@laas.fr>
+ * @author Ayoub Farah Hassan <ayoub.farah-hassan@laas.fr>
  */
 
 //--------------OWNTECH APIs----------------------------------
@@ -32,32 +33,17 @@
 #include "TaskAPI.h"
 #include "TwistAPI.h"
 #include "SpinAPI.h"
-#include "opalib_control_pid.h"
-
-#include "zephyr/console/console.h"
 
 //--------------SETUP FUNCTIONS DECLARATION-------------------
 void setup_routine(); // Setups the hardware and software of the system
 
 //--------------LOOP FUNCTIONS DECLARATION--------------------
-void loop_communication_task(); // code to be executed in the slow communication task
-void loop_application_task();   // Code to be executed in the background task
+void loop_background_task();   // Code to be executed in the background task
 void loop_critical_task();     // Code to be executed in real time in the critical task
 
 //--------------USER VARIABLES DECLARATIONS-------------------
-uint8_t received_serial_char;
 
-float32_t duty_cycle = 0.3;
-
-//---------------------------------------------------------------
-
-enum serial_interface_menu_mode // LIST OF POSSIBLE MODES FOR THE OWNTECH CONVERTER
-{
-    IDLEMODE = 0,
-    POWERMODE
-};
-
-uint8_t mode = IDLEMODE;
+static uint32_t incremental_value;
 
 //--------------SETUP FUNCTIONS-------------------------------
 
@@ -70,68 +56,34 @@ uint8_t mode = IDLEMODE;
 void setup_routine()
 {
     // Setup the hardware first
-    spin.version.setBoardVersion(TWIST_v_1_1_2);
-    twist.setVersion(shield_TWIST_V1_2);
+    spin.version.setBoardVersion(SPIN_v_1_0);
 
-    spin.pwm.setFrequency(200000); // Set frequency of pwm
-    spin.pwm.initUnit(PWMA); // timer initialization
-
-    spin.pwm.startDualOutput(PWMA); // Start PWM
+    spin.timer.startLogTimer4IncrementalEncoder();
 
     // Then declare tasks
-    uint32_t app_task_number = task.createBackground(loop_application_task);
-    uint32_t com_task_number = task.createBackground(loop_communication_task);
-    task.createCritical(loop_critical_task, 100); // Uncomment if you use the critical task
+    uint32_t background_task_number = task.createBackground(loop_background_task);
+    //task.createCritical(loop_critical_task, 500); // Uncomment if you use the critical task
 
     // Finally, start tasks
-    task.startBackground(app_task_number);
-    task.startBackground(com_task_number);
-    task.startCritical(); // Uncomment if you use the critical task
+    task.startBackground(background_task_number);
+    //task.startCritical(); // Uncomment if you use the critical task
 }
 
 //--------------LOOP FUNCTIONS--------------------------------
-
-void loop_communication_task()
-{
-    while (1)
-    {
-        received_serial_char = console_getchar();
-        switch (received_serial_char)
-        {
-        case 'h':
-            //----------SERIAL INTERFACE MENU-----------------------
-            printk(" ________________________________________\n");
-            printk("|     ------- MENU ---------             |\n");
-            printk("|     press u : duty cycle UP            |\n");
-            printk("|     press d : duty cycle DOWN          |\n");
-            printk("|________________________________________|\n\n");
-            //------------------------------------------------------
-            break;
-        case 'u':
-            duty_cycle += 0.05;
-            break;
-        case 'd':
-            duty_cycle -= 0.05;
-            break;
-        default:
-            break;
-        }
-    }
-}
 
 /**
  * This is the code loop of the background task
  * It is executed second as defined by it suspend task in its last line.
  * You can use it to execute slow code such as state-machines.
  */
-void loop_application_task()
+void loop_background_task()
 {
+    incremental_value = spin.timer.getTimer4IncrementalEncoderValue();
     // Task content
-    printk("%f\n", duty_cycle);
+    printk(" %u \n", incremental_value);
 
     // Pause between two runs of the task
-    task.suspendBackgroundMs(1000);
-
+    task.suspendBackgroundMs(100);
 }
 
 /**
@@ -142,7 +94,7 @@ void loop_application_task()
  */
 void loop_critical_task()
 {
-        spin.pwm.setDutyCycle(PWMA, duty_cycle);
+
 }
 
 /**
