+++ TWIST/DC_DC/buck_current_mode/README.md
@@ -1,4 +1,4 @@
-# Buck voltage mode
+# Buck current mode
 
 ## Overview
