+++ TWIST/DC_DC/buck_voltage_mode/main.cpp
@@ -25,6 +25,7 @@
  *
  * @author Clément Foucher <clement.foucher@laas.fr>
  * @author Luiz Villa <luiz.villa@laas.fr>
+ * @author Ayoub Farah Hassan <ayoub.farah-hassan@laas.fr>
  */
 
 //--------------OWNTECH APIs----------------------------------
@@ -162,27 +163,22 @@ void loop_communication_task()
  */
 void loop_application_task()
 {
-    while (1)
+    if (mode == IDLEMODE)
     {
-
-        if (mode == IDLEMODE)
-        {
-            spin.led.turnOff();
-        }
-        else if (mode == POWERMODE)
-        {
-            spin.led.turnOn();
-
-            printk("%f:", I1_low_value);
-            printk("%f:", V1_low_value);
-            printk("%f:", I2_low_value);
-            printk("%f:", V2_low_value);
-            printk("%f:", I_high);
-            printk("%f\n", V_high);
-        }
-        k_msleep(100);
+        spin.led.turnOff();
     }
+    else if (mode == POWERMODE)
+    {
+        spin.led.turnOn();
 
+        printk("%f:", I1_low_value);
+        printk("%f:", V1_low_value);
+        printk("%f:", I2_low_value);
+        printk("%f:", V2_low_value);
+        printk("%f:", I_high);
+        printk("%f\n", V_high);
+    }
+    task.suspendBackgroundMs(100);
 }
 
 /**
