+++ TWIST/Microgrid/DC_client_server/app.conf
@@ -1,0 +1,1 @@
+CONFIG_OWNTECH_COMMUNICATION=y
