+++ TWIST/Microgrid/DC_client_server/main.cpp
@@ -25,6 +25,8 @@
  *
  * @author Clément Foucher <clement.foucher@laas.fr>
  * @author Luiz Villa <luiz.villa@laas.fr>
+ * @author Antoine Boche <antoine.boche@laas.fr>
+ * @author Ayoub Farah Hassan <ayoub.farah-hassan@laas.fr>
  */
 
 //--------------OWNTECH APIs----------------------------------
@@ -72,7 +74,9 @@ static float32_t I2_low_value;
 
 // reference voltage/current
 float32_t duty_cycle = 0.5;
+#ifdef MASTER
 static float32_t Vref = 12.0;
+#endif
 static float32_t Iref;
 static float32_t PeakRef_Raw;
 int count = 0;
@@ -186,25 +190,20 @@ void loop_communication_task()
  */
 void loop_application_task()
 {
-    while (1)
+    if (mode == IDLEMODE)
     {
-
-        if (mode == IDLEMODE)
-        {
-            spin.led.turnOff();
-        }
-        else if (mode == POWERMODE)
-        {
-            spin.led.turnOn();
-
-            printk("%f:", I1_low_value);
-            printk("%f:", V1_low_value);
-            printk("%f:", I2_low_value);
-            printk("%f\n", V2_low_value);
-        }
-        k_msleep(100);
+        spin.led.turnOff();
     }
+    else if (mode == POWERMODE)
+    {
+        spin.led.turnOn();
 
+        printk("%f:", I1_low_value);
+        printk("%f:", V1_low_value);
+        printk("%f:", I2_low_value);
+        printk("%f\n", V2_low_value);
+    }
+    task.suspendBackgroundMs(100);
 }
 
 /**
