+++ TWIST/Microgrid/DC_peer_to_peer/app.conf
@@ -1,0 +1,1 @@
+CONFIG_OWNTECH_COMMUNICATION=y
